requirements.txt
langchain==0.0.205
openai==0.27.7
pinecone_client==2.2.1
sentence_transformers==2.2.2
streamlit_chat==0.0.2.2
streamlit==1.22.0
